# api/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateSubscriptionRequest:
    description: Defines the request body for creating a new subscription.
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        maxLength: 15
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.ErrorResponse:
    description: Defines the generic error
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  models.ListSubscriptionsResponse:
    description: Defines the API response structure for a ListSubscriptionRequest.
    properties:
      meta:
        $ref: '#/definitions/models.PaginationMeta'
      subscriptions:
        items:
          $ref: '#/definitions/models.SubscriptionResponse'
        type: array
    type: object
  models.PaginationMeta:
    description: Defines pagination metadata for response for ListSubscriptionResponse
    properties:
      limit:
        type: integer
      offset:
        type: integer
      order:
        type: string
      sort_by:
        type: string
      total:
        type: integer
    type: object
  models.SubscriptionResponse:
    description: Defines the API response structure for a subscription.
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_id:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  models.UpdateSubscriptionRequest:
    description: Defines the request body for updating a subscription.
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        maxLength: 15
        type: string
      start_date:
        type: string
    type: object
  models.UserSubscriptionSummaryResponse:
    description: Defines the structure of the API response for the /summary endpoint.
    properties:
      service_name:
        type: string
      total_amount:
        type: integer
      total_months:
        type: integer
      unit_price:
        type: integer
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing user subscriptions
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Retrieve paginated list of subscriptions with optional sorting
      parameters:
      - default: 10
        description: Maximum number of items to return
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        minimum: 0
        name: offset
        type: integer
      - default: id
        description: Field to sort by
        enum:
        - id
        - user_id
        - service_name
        - price
        - start_date
        - end_date
        in: query
        name: sort_by
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ListSubscriptionsResponse'
        "400":
          description: Bad Request - Invalid query parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List subscriptions with pagination
      tags:
      - Subscriptions
    post:
      consumes:
      - application/json
      description: Create a subscription for a user
      parameters:
      - description: Subscription payload
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new subscription
      tags:
      - Subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Subscription successfully deleted
        "400":
          description: Bad Request - Invalid subscription ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found - Subscription does not exist
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete subscription
      tags:
      - Subscriptions
    get:
      consumes:
      - application/json
      description: Retrieve a subscription using its ID
      parameters:
      - description: Subscription ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad Request - Invalid subscription ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found - Subscription does not exist
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - Subscriptions
    put:
      consumes:
      - application/json
      description: Partially update subscription fields by ID (only provided fields
        are modified)
      parameters:
      - description: Subscription ID
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      - description: Update payload (partial update)
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SubscriptionResponse'
        "400":
          description: Bad Request - Invalid input or validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found - Subscription does not exist
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update subscription
      tags:
      - Subscriptions
  /subscriptions/summary:
    get:
      consumes:
      - application/json
      description: Calculate subscription statistics including total cost, unique
        months, and count for a user
      parameters:
      - description: User UUID
        format: uuid
        in: query
        name: user_id
        required: true
        type: string
      - description: Filter by service name
        in: query
        name: service_name
        required: true
        type: string
      - description: Start date (MM-YYYY)
        in: query
        name: from
        type: string
      - description: End date (MM-YYYY)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserSubscriptionSummaryResponse'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get user subscription summary
      tags:
      - Subscriptions
swagger: "2.0"